refactor: Restructure tests
